kernel: add spinlock wrappers, acquire/release
Note:
spin_lock() and spin_unlock() have been modified to
accept pointer to lock as argument. Earlier instead of `ecx`
register `ebx` was used, but it was somehow corrupting callers
stack. Later found that, `ebx` is `callee` saved register,
hence before using it needs to be pused to stack for later restoration.

Signed-off-by: Mahavir Jain <mahavirpj@gmail.com>