swtch: context switching code check if both operands are same
If current task is trying to switch to itself then, that will
not work as, context would be pushed to its stack, but restored
from older stack pointer reference from second argument.

Ideally this should be handled in scheduler code itself, but
this is just to be on safer side.

Signed-off-by: Mahavir Jain <mahavirpj@gmail.com>